Moving code that's in common over to kubernetes_worker, where it is currently used. We don't have the docker layer included on kubernetes-master and it failed to deploy with this change in common.
