+++ cluster/juju/layers/kubernetes-common/lib/charms/layer/kubernetes_common.py
@@ -21,7 +21,6 @@ import subprocess
 import hashlib
 import json
 import random
-import shutil
 
 from pathlib import Path
 from subprocess import check_output, check_call
@@ -30,9 +29,7 @@ from shlex import split
 from subprocess import CalledProcessError
 from charmhelpers.core import hookenv, unitdata
 from charmhelpers.core import host
-from charms import layer
-from charms.docker import DockerOpts
-from charms.reactive import endpoint_from_flag, is_state, set_flag
+from charms.reactive import endpoint_from_flag, is_state
 from time import sleep
 
 db = unitdata.kv()
@@ -486,61 +483,3 @@ def configure_kube_proxy(configure_prefix, api_servers, cluster_cidr):
 
     configure_kubernetes_service(configure_prefix, 'kube-proxy',
                                  kube_proxy_opts, 'proxy-extra-args')
-
-
-def manage_docker_opts(opts, remove=False):
-    '''Add or remove docker daemon options.
-
-    Options here will be merged with configured docker-opts when layer-docker
-    processes a daemon restart.
-
-    :param: dict opts: option keys/values; use None value if the key is a flag
-    :param: bool remove: True to remove the options; False to add them
-    '''
-    docker_opts = DockerOpts()
-    for k, v in opts.items():
-        # Always remove existing option
-        if docker_opts.exists(k):
-            docker_opts.pop(k)
-        if not remove:
-            docker_opts.add(k, v)
-    hookenv.log('DockerOpts daemon options changed. Requesting a restart.')
-    # State will be removed by layer-docker after restart
-    set_flag('docker.restart')
-
-
-def manage_registry_certs(subdir, remove=False):
-    '''Add or remove TLS data for a specific registry.
-
-    When present, the docker client will use certificates when communicating
-    with a specific registry.
-
-    :param: str subdir: subdirectory to store the client certificates
-    :param: bool remove: True to remove cert data; False to add it
-    '''
-    cert_dir = '/etc/docker/certs.d/{}'.format(subdir)
-
-    if remove:
-        if os.path.isdir(cert_dir):
-            hookenv.log('Disabling registry TLS: {}.'.format(cert_dir))
-            shutil.rmtree(cert_dir)
-    else:
-        tls_options = layer.options('tls-client')
-        client_cert_path = tls_options.get('client_certificate_path')
-        client_key_path = tls_options.get('client_key_path')
-
-        os.makedirs(cert_dir, exist_ok=True)
-        client_tls = {
-            client_cert_path: '{}/client.cert'.format(cert_dir),
-            client_key_path: '{}/client.key'.format(cert_dir),
-        }
-        for f, link in client_tls.items():
-            if os.path.isfile(f):
-                try:
-                    os.remove(link)
-                except FileNotFoundError:
-                    pass
-                hookenv.log('Creating registry TLS link: {}.'.format(link))
-                os.symlink(f, link)
-            else:
-                hookenv.log('Missing TLS file for registry: {}.'.format(f))
